Time: 39 ms (96.74%), Space: 53.8 MB (35.12%) - LeetHub

+++ 0041-first-missing-positive/0041-first-missing-positive.cpp
@@ -4,7 +4,7 @@ public:
         int n = nums.size();
         int i = 0;
         while (i < n){
-            int ind = nums[i]-1;
+            long ind = (long)nums[i]-1;
             bool isValid = (ind >= 0 && ind < n);
             if (isValid && nums[ind] != nums[i]){
                 swap(nums[ind], nums[i]);
